+++ utide/band_average.py
